Style Cleaned up and organized.  Font issues corrected.  Themes simplified
