+++ src/assets/styles/scss/app.module.scss
@@ -346,50 +346,6 @@ body.light-mode .video-bg:before {
  overflow: hidden;
 }
 
-.left-side {
- flex-basis: 240px;
- border-right: 1px solid var(--border-color);
- padding: 26px;
- overflow: auto;
- flex-shrink: 0;
- @media screen and (max-width: 945px) {
-  display: none;
- }
-}
-
-.side-wrapper + .side-wrapper {
- margin-top: 20px;
-}
-
-.side-title {
- color: var(--inactive-color);
- margin-bottom: 14px;
-}
-
-.side-menu {
- display: flex;
- flex-direction: column;
- white-space: nowrap;
- a {
-  text-decoration: none;
-  color: var(--theme-color);
-  display: flex;
-  align-items: center;
-  font-weight: 400;
-  padding: 10px;
-  font-size: 14px;
-  border-radius: 6px;
-  transition: 0.3s;
-  &:hover {
-   background-color: var(--hover-menu-bg);
-  }
- }
- svg {
-  width: 16px;
-  margin-right: 8px;
- }
-}
-
 .updates {
  position: relative;
  top: 0;
@@ -448,7 +404,7 @@ body.light-mode .video-bg:before {
   align-items: center;
   width: 100%;
   justify-content: space-between;
-  filter: drop-shadow( 10px 20px 15px #000 );
+  filter: drop-shadow( 10px 20px 15px rgb(138, 4, 4) );
   background-image: 
    linear-gradient(
     to right top,
