+++ src/assets/styles/themes/reset.tsx
@@ -1,0 +1,10 @@
+const base = {
+  styles: {
+    global: () => ({
+      body: {
+        bg: ''
+      }
+    })
+  }
+};
+export default base;
