+++ src/features/nav/left/menu.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import styles from 'styles/app.module.scss';
+import styles from '@scss/side.module.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { solid } from '@fortawesome/fontawesome-svg-core/import.macro';
