+++ src/layouts/Left.tsx
@@ -1,4 +1,4 @@
-import styles from '@scss/app.module.scss';
+import styles from '@scss/side.module.scss';
 import NavLeft from '@features/nav/left/menu';
 
 export default function layoutLeft() {
