+++ src/styles/app.module.scss
@@ -437,7 +437,6 @@ body.light-mode .video-bg:before {
  color: var(--theme-color);
  padding: 20px 40px;
  height: 100%;
- overflow: auto;
  background-color: var(--theme-bg-color);
  @media screen and (max-width: 510px) {
   padding: 20px;
